Extract App component to separate file
